// legacy/v1/frontend/src/api/generated/models/sptransaction-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * SP取引の種類
 * @export
 * @enum {string}
 */

export const SPTransactionType = {
    DailyRecovery: 'daily_recovery',
    Purchase: 'purchase',
    Achievement: 'achievement',
    EventReward: 'event_reward',
    LogResult: 'log_result',
    LoginBonus: 'login_bonus',
    Refund: 'refund',
    FreeAction: 'free_action',
    LogDispatch: 'log_dispatch',
    LogEnhancement: 'log_enhancement',
    MemoryInheritance: 'memory_inheritance',
    SystemFunction: 'system_function',
    Movement: 'movement',
    Exploration: 'exploration',
    Adjustment: 'adjustment',
    Migration: 'migration',
    AdminGrant: 'admin_grant',
    AdminDeduct: 'admin_deduct'
} as const;

export type SPTransactionType = typeof SPTransactionType[keyof typeof SPTransactionType];